WEBAPP-224: Extract to LocationHeaderEventsNavigationItem
