+++ src/modules/layout/containers/LocationHeader.js
@@ -8,14 +8,14 @@ import locationIcon from '../assets/location-icon.svg'
 import languageIcon from '../assets/language-icon.svg'
 import LocationModel from 'modules/endpoint/models/LocationModel'
 import Header from 'modules/layout/components/Header'
-import HeaderNavigationItem from './HeaderNavigationItem'
+import HeaderNavigationItem from '../components/HeaderNavigationItem'
 import HeaderActionItem from '../HeaderActionItem'
 import SearchPage from 'routes/search/containers/SearchPage'
 import LandingPage from 'routes/landing/containers/LandingPage'
 import CategoriesPage from 'routes/categories/containers/CategoriesPage'
 import EventsPage from 'routes/events/containers/EventsPage'
 import ExtrasPage from 'routes/extras/containers/ExtrasPage'
-import withFetcher from '../../endpoint/hocs/withFetcher'
+import LocationHeaderEventsNavigationItem from './LocationHeaderEventsNavigationItem'
 
 class LocationHeader extends React.Component {
   static propTypes = {
@@ -68,15 +68,7 @@ class LocationHeader extends React.Component {
       text={t('categories')}
       active />
 
-    const EventsInactive = props => <HeaderNavigationItem {...props}
-                                                          href={matchRoute(EventsPage).stringify(currentParams)}
-                                                          selected={isEventsSelected()}
-                                                          text={t('news')}
-                                                          tooltip={t('noNews')} active />
-    const EventsFetched = props => <HeaderNavigationItem active={props.events.length > 0} {...props} />
-    const EventsNavigationItem = withFetcher('events', EventsInactive, EventsInactive)(EventsFetched)
-
-    const events = isEventsEnabled() && <EventsNavigationItem
+    const events = isEventsEnabled() && <LocationHeaderEventsNavigationItem
       href={matchRoute(EventsPage).stringify(currentParams)}
       selected={isEventsSelected()}
       text={t('news')}
